+++ saleor/graphql/product/types.py
@@ -200,6 +200,17 @@ class ProductVariant(CountableDjangoObjectType):
     def resolve_images(self, info):
         return self.images.all()
 
+    @classmethod
+    def get_node(cls, info, id):
+        user = info.context.user
+        visible_products = models.Product.objects.visible_to_user(
+            user).values_list('pk', flat=True)
+        try:
+            return cls._meta.model.objects.filter(
+                product__id__in=visible_products).get(pk=id)
+        except cls._meta.model.DoesNotExist:
+            return None
+
 
 class ProductAvailability(graphene.ObjectType):
     available = graphene.Boolean()
@@ -344,6 +355,17 @@ class Product(CountableDjangoObjectType):
     def resolve_available_on(self, info):
         return self.publication_date
 
+    @classmethod
+    def get_node(cls, info, id):
+        if info.context:
+            user = info.context.user
+            try:
+                return cls._meta.model.objects.visible_to_user(
+                    user).get(pk=id)
+            except cls._meta.model.DoesNotExist:
+                return None
+        return None
+
 
 def prefetch_products(info, *args, **kwargs):
     """Prefetch products visible to the current user.
@@ -424,6 +446,17 @@ class Collection(CountableDjangoObjectType):
             return self.prefetched_products
         qs = self.products.visible_to_user(info.context.user)
         return gql_optimizer.query(qs, info)
+
+    @classmethod
+    def get_node(cls, info, id):
+        if info.context:
+            user = info.context.user
+            try:
+                return cls._meta.model.objects.visible_to_user(
+                    user).get(pk=id)
+            except cls._meta.model.DoesNotExist:
+                return None
+        return None
 
     def resolve_published_date(self, info):
         return self.publication_date